Move endpoint to supplier path
